Enum mem4_common::WsMessage

pub enum WsMessage {
    Dummy {
        dummy: String,
    },
    RequestWsUid {
        test: String,
    },
    ResponseWsUid {
        your_ws_uid: usize,
    },
    WantToPlay {
        my_ws_uid: usize,
        asked_folder_name: String,
    },
    PlayAccept {
        my_ws_uid: usize,
        players: String,
    },
    GameDataInit {
        card_grid_data: String,
        game_config: String,
        players: String,
    },
    PlayerClick1Card {
        my_ws_uid: usize,
        players: String,
        card_index: usize,
        game_status: GameStatus,
    },
    PlayerClick2Card {
        my_ws_uid: usize,
        players: String,
        card_index: usize,
        game_status: GameStatus,
    },
    PlayAgain {
        my_ws_uid: usize,
        players: String,
    },
    TakeTurnEnd {
        my_ws_uid: usize,
        players: String,
    },
    RequestGameConfig {
        filename: String,
    },
    ResponseGameConfigJson {
        json: String,
    },
}

WsMessage enum for WebSocket

Variants
Dummy

Dummy

Fields of Dummy

dummy: String

anything

RequestWsUid

Request WebSocket Uid - first message to WebSocket server

Fields of RequestWsUid

test: String

anything

ResponseWsUid

response from WebSocket server for first message

Fields of ResponseWsUid

your_ws_uid: usize

WebSocket Uid

WantToPlay

want to play

Fields of WantToPlay

my_ws_uid: usize

ws client instance unique id. To not listen the echo to yourself.

asked_folder_name: String

content folder name

PlayAccept

accept play

Fields of PlayAccept

my_ws_uid: usize

ws client instance unique id. To not listen the echo to yourself.

players: String

json of vector of players

GameDataInit

player1 initialize the game data and sends it to all players I will send json string to not confuse the server with vectors

Fields of GameDataInit

card_grid_data: String

act is the action to take on the receiver

game_config: String

json of game_config

players: String

json of vector of players

PlayerClick1Card

player click

Fields of PlayerClick1Card

my_ws_uid: usize

this identifies the smartphone, but not the player-in-turn

players: String

all players

card_index: usize

card index

game_status: GameStatus

game status PlayerBefore1Card or PlayerBefore2Card

PlayerClick2Card

player click

Fields of PlayerClick2Card

my_ws_uid: usize

this identifies the smartphone, but not the player-in-turn

players: String

all players

card_index: usize

card index

game_status: GameStatus

game status PlayerBefore1Card or PlayerBefore2Card

PlayAgain

Play Again

Fields of PlayAgain

my_ws_uid: usize

this identifies the smartphone, but not the player-in-turn

players: String

all players

TakeTurnEnd

player change

Fields of TakeTurnEnd

my_ws_uid: usize

ws client instance unique id. To not listen the echo to yourself.

players: String

all players

RequestGameConfig

Request the game_config from the WebSocket server

Fields of RequestGameConfig

filename: String

the file with the game_config

ResponseGameConfigJson

Receive the game_config from the WebSocket server

Fields of ResponseGameConfigJson

json: String

the game_config from the server